+++ .gitignore
@@ -1,3 +1,5 @@
+# Cypress screenshots
+cypress/screenshots/
 /node_modules
 .next/
 .env
