+++ package.json
@@ -5,7 +5,8 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "test": "dotenv -e .env.test -- npx prisma migrate deploy && dotenv -e .env.test -- cypress run"
+    "test": "npx dotenv-cli -e .env.test -- npx prisma migrate deploy && npx dotenv-cli -e .env.test -- cypress run",
+    "dev:test": "npx dotenv-cli -e .env.test -- next dev -p 3001"
   },
   "dependencies": {
     "@next/font": "^14.2.15",
@@ -19,7 +20,8 @@
     "postcss": "^8.5.2",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
-    "tailwindcss": "^4.0.6"
+    "tailwindcss": "^4.0.6",
+    "zod": "^4.1.12"
   },
   "devDependencies": {
     "@types/bluebird": "^3.5.42",
