+++ pages/api/users.ts
@@ -1,5 +1,7 @@
+
 import { PrismaClient } from "@prisma/client";
 import { Request, Response } from "../../types";
+import { z } from "zod";
 
 const prisma = new PrismaClient();
 
@@ -10,8 +12,18 @@ export default async function handler(req: Request, res: Response) {
   }
 
   if (req.method === "POST") {
-    const { name, email } = req.body;
-    const newUser = await prisma.user.create({ data: { name, email } });
+    const userSchema = z.object({
+      name: z.string(),
+      email: z.string().email(),
+      password: z.string(),
+    });
+
+    const parseResult = userSchema.safeParse(req.body);
+    if (!parseResult.success) {
+      return res.status(400).json({ error: parseResult.error.flatten() });
+    }
+
+    const newUser = await prisma.user.create({ data: parseResult.data });
     return res.status(201).json(newUser);
   }
 }
