+++ types/next-auth.d.ts
@@ -1,0 +1,15 @@
+import NextAuth from "next-auth";
+
+declare module "next-auth" {
+  interface Session {
+    user: {
+      id: string;
+      name?: string | null;
+      email?: string | null;
+      image?: string | null;
+      active?: boolean | null;
+    };
+  }
+}
+
+
